getBaseNamespace rewritten for robustness.
model.getNsPrefixURI("") is unreliable, as sometimes xml:base is not present and when xml:base is after other namespaces in rdf:RDF, this call returns null. Added additional methods in case the original fails and a check for symptoms of inconsistencies.

All three methods are essentially hacks without support from the specifications. IMO "ontologies" should be identified by hashes as far as Alignator is concerned (also, getOntology(type) should probably return a list of ontologies).